Move api tests to subdirectory
